So i've added repo validation to the modify tables functions

+++ allowList/modifyTables/allowlist_modifyTables.py
@@ -10,6 +10,7 @@ Last Modified by : James Miller (z5257531)
 # External Imports
 import boto3
 from boto3.dynamodb.conditions import Key
+import botocore
 from botocore.exceptions import ClientError
 import time
 
@@ -30,7 +31,7 @@ def create_organization_table(org_id, dynamodb=None):
 		dynamodb (dynamodb service resource, optional): DynamoDB connection. Defaults to None, which will uses the localhost:8000 instead of any region
 
 	Raises:
-		ValueError: if the repo_id is None, or a repo already exists for the given ID
+		ValueError: if the org_id is None, or a table already exists for the given ID
 
 	Returns:
 		True: If a new table is successfully created the function will return true
@@ -38,57 +39,92 @@ def create_organization_table(org_id, dynamodb=None):
 	dynamodb = boto3.resource('dynamodb', endpoint_url="http://localhost:8000") if not dynamodb else dynamodb
 
 	table_name = f'{tablename_prefix}_{org_id}'
-	
-	dynamodb.create_table(
-		TableName=table_name,
-		KeySchema=[
-			{
-				'AttributeName': 'repo_id',
-				'KeyType': 'HASH' # Partition key
-			}, 
-			{
-				'AttributeName': 'whitelist_term',
-				'KeyType': 'RANGE'  # Sort key
-			}
-		],
-		LocalSecondaryIndexes=[
-			{
-				'IndexName':"LatestAdditions",
-				'KeySchema':[
-					{
-						'AttributeName': 'repo_id',
-						'KeyType': 'HASH' # Partition key
-					}, 
-					{
-						'AttributeName': 'time_added',
-						'KeyType': 'RANGE'  # Sort key
-					}
-				],
-				'Projection':{
-					'ProjectionType':'ALL'
+	try:
+		dynamodb.create_table(
+			TableName=table_name,
+			KeySchema=[
+				{
+					'AttributeName': 'repo_id',
+					'KeyType': 'HASH' # Partition key
+				}, 
+				{
+					'AttributeName': 'whitelist_term',
+					'KeyType': 'RANGE'  # Sort key
 				}
+			],
+			LocalSecondaryIndexes=[
+				{
+					'IndexName':"LatestAdditions",
+					'KeySchema':[
+						{
+							'AttributeName': 'repo_id',
+							'KeyType': 'HASH' # Partition key
+						}, 
+						{
+							'AttributeName': 'time_added',
+							'KeyType': 'RANGE'  # Sort key
+						}
+					],
+					'Projection':{
+						'ProjectionType':'ALL'
+					}
+				}
+			],
+			AttributeDefinitions=[
+				{
+					'AttributeName': 'repo_id',
+					'AttributeType': 'S' # string
+				}, 
+				{
+					'AttributeName': 'whitelist_term',
+					'AttributeType': 'S' # string
+				},
+				{
+					'AttributeName': 'time_added',
+					'AttributeType': 'N' # Epoch Time
+				}
+			],
+			ProvisionedThroughput={
+				'ReadCapacityUnits': 5,
+				'WriteCapacityUnits': 5
 			}
-		],
-		AttributeDefinitions=[
-			{
-				'AttributeName': 'repo_id',
-				'AttributeType': 'S' # string
-			}, 
-			{
-				'AttributeName': 'whitelist_term',
-				'AttributeType': 'S' # string
-			},
-			{
-				'AttributeName': 'time_added',
-				'AttributeType': 'N' # Epoch Time
-			}
-		],
-		ProvisionedThroughput={
-			'ReadCapacityUnits': 5,
-			'WriteCapacityUnits': 5
-		}
-	)
-	return True
+		)
+	except Exception as e:
+		if str(e) == "An error occurred (ResourceInUseException) when calling the CreateTable operation: Cannot create preexisting table":
+			return False
+		else:
+			raise e
+	else:
+		return True
+
+@validateID
+def setup_new_repo(org_id, repo_id, dynamodb=None):
+	dynamodb = boto3.resource('dynamodb', endpoint_url="http://localhost:8000") if not dynamodb else dynamodb
+	table_name = f'{tablename_prefix}_{org_id}'
+	repo_id = f'repo_{repo_id}'
+
+	query_params = {
+		'KeyConditionExpression':  Key('repo_id').eq(repo_id), 
+		'ProjectionExpression': "whitelist_term",
+		'ConsistentRead': True
+	}
+	table = dynamodb.Table(table_name)
+	response = table.query( **query_params)
+
+	if len(response['Items']):
+		raise ValueError("Repo with given ID already initialized")
+
+	creationTime = int(time.time())
+	item = {
+		'repo_id' : repo_id,
+		'whitelist_term': "_initialization_",
+		'time_added': creationTime,
+		'info': {}
+	}
+	table = dynamodb.Table(table_name)
+	response = table.put_item(Item=item)
+
+	return response['ResponseMetadata']['HTTPStatusCode'] == 200
 
 @validateID
 def read_repo(org_id, repo_id, dynamodb=None):
@@ -132,6 +168,11 @@ def read_repo(org_id, repo_id, dynamodb=None):
 		# There is a possability the first pass won't leave any terms left - this is to catch that exclusive case
 		pass
 
+	if not to_return:
+		raise ValueError('ValueError: Repo has not been initialized')
+
+	to_return.remove('_initialization_')
+
 	return to_return
 
 @validateID
@@ -170,14 +211,18 @@ def read_repo_with_info(org_id, repo_id, dynamodb=None):
 				ExclusiveStartKey = response['LastEvaluatedKey']
 			)
 			for item in response['Items']:
-				to_return.append(item['whitelist_term'])
+				to_return.append({'term':item['whitelist_term'], 'info':item['info']})
 			
 	except KeyError as e:
 		# There is a possability the first pass won't leave any terms left - this is to catch that exclusive case
 		pass
 
-	return to_return
+	if not to_return:
+		raise ValueError('ValueError: Repo has not been initialized')
+	
+	to_return.remove({'term':'_initialization_', 'info':{}})
 
+	return to_return
 
 @validateID
 def insert_new_term(org_id, repo_id, new_term, other_info=None, dynamodb=None):
@@ -200,13 +245,25 @@ def insert_new_term(org_id, repo_id, new_term, other_info=None, dynamodb=None):
 	repo_id = f'repo_{repo_id}'
 	creationTime = int(time.time())
 
+	table = dynamodb.Table(table_name)
+
+	query_params = {
+		'KeyConditionExpression':  Key('repo_id').eq(repo_id), 
+		'ProjectionExpression': "whitelist_term",
+		'ConsistentRead': True
+	}
+	
+	response = table.query( **query_params)
+
+	if not len(response['Items']):
+		raise ValueError('ValueError: Repo has not been initialized')
+
 	item = {
 		'repo_id' : repo_id,
 		'whitelist_term': new_term,
 		'time_added': creationTime,
 		'info': other_info
 	}
-	table = dynamodb.Table(table_name)
 	response = table.put_item(Item=item)
 
 	return response
@@ -215,6 +272,8 @@ def insert_new_term(org_id, repo_id, new_term, other_info=None, dynamodb=None):
 def insert_new_terms(org_id, repo_id, new_terms, dynamodb=None):
 	"""
 	Same as insert_new_term, except for multiple iteratable terms
+
+	Throws: Resource in use Exception
 
 	Args:
 		org_id (int/str): id of the organization to add term to
@@ -228,6 +287,17 @@ def insert_new_terms(org_id, repo_id, new_terms, dynamodb=None):
 
 	table = dynamodb.Table(table_name)
 
+	query_params = {
+		'KeyConditionExpression':  Key('repo_id').eq(repo_id), 
+		'ProjectionExpression': "whitelist_term",
+		'ConsistentRead': True
+	}
+	
+	response = table.query( **query_params)
+
+	if not len(response['Items']):
+		raise ValueError('ValueError: Repo has not been initialized')
+	
 	with table.batch_writer() as batch:
 		for term in new_terms:
 			creationTime = int(time.time())
@@ -260,6 +330,17 @@ def insert_new_terms_with_info(org_id, repo_id, new_terms, dynamodb=None):
 	table_name = f'{tablename_prefix}_{org_id}'
 	repo_id = f'repo_{repo_id}'
 	table = dynamodb.Table(table_name)
+
+	query_params = {
+		'KeyConditionExpression':  Key('repo_id').eq(repo_id), 
+		'ProjectionExpression': "whitelist_term",
+		'ConsistentRead': True
+	}
+	
+	response = table.query( **query_params)
+
+	if not len(response['Items']):
+		raise ValueError('ValueError: Repo has not been initialized')
 
 	with table.batch_writer() as batch:
 		for term in new_terms:
@@ -297,6 +378,18 @@ def delete_term(org_id, repo_id, term, dynamodb=None):
 
 	try:
 		table = dynamodb.Table(table_name)
+
+		query_params = {
+			'KeyConditionExpression':  Key('repo_id').eq(repo_id), 
+			'ProjectionExpression': "whitelist_term",
+			'ConsistentRead': True
+		}
+		
+		response = table.query( **query_params)
+
+		if not len(response['Items']):
+			raise ValueError('ValueError: Repo has not been initialized')
+
 		response = table.delete_item(
 			Key = {
 				'repo_id' : repo_id,
@@ -306,6 +399,8 @@ def delete_term(org_id, repo_id, term, dynamodb=None):
 	except ClientError as e:
 		# TODO
 		raise ValueError("There was a bad paramater")
+	except ValueError as e:
+		raise e
 	else:
 		return response
 
@@ -334,6 +429,10 @@ def delete_repo(org_id, repo_id, dynamodb=None):
 
 	table = dynamodb.Table(table_name)
 	response = table.query( **query_params)
+
+	if not len(response['Items']):
+		raise ValueError('ValueError: Repo has not been initialized')
+	
 	deleted_items += len(response['Items'])
 	with table.batch_writer() as batch:
 			for item in response['Items']:
